+++ src/useful/libs/Useful.Assets/TrueTypeFontEntry.cs
@@ -1,0 +1,12 @@
+// 'Useful Libraries' - Andy Hawkins 2023-2026.
+
+namespace Useful.Assets;
+
+//// JSON serializable
+
+public class TrueTypeFontEntry
+{
+    public string File { get; set; } = string.Empty;
+
+    public int PointSize { get; set; }
+}